refactor(musefs-format): split mp4 box parsing from layout synthesis
Box parsing/reading and layout synthesis lived side by side in one file. Move
the synthesis half (`boxed`..`synthesize_layout`) into `mp4/synth.rs`, keeping
the parsing/reading API in `mp4.rs`, and extract the inline test module to
`mp4/tests.rs`. mp4.rs drops 2582 -> ~590 lines.

The public API is preserved: `read_*`/`locate_audio`/`read_structure*` stay in
mp4.rs and `synthesize_layout` is re-exported, so external consumers
(musefs-core, the out-of-workspace fuzz crate) are unaffected — verified with
`cargo +nightly fuzz build mp4`. The four private synth helpers the tests use
(`boxed`, `build_udta`, `freeform_binary_prefix`, `patch_chunk_offsets`) are
bumped to `pub(super)`; `synth.rs` uses explicit imports (pedantic
`wildcard_imports` forbids `use super::*` outside test modules).

Addresses #445 (concern split) and #444 (test extraction) for mp4.
